moved global back handling to App.tsx
